+++ src/languages/pygrep/mod.rs
@@ -1,0 +1,4 @@
+#[allow(clippy::module_inception)]
+mod pygrep;
+
+pub(crate) use pygrep::Pygrep;
