+++ src/languages/python/mod.rs
@@ -4,5 +4,6 @@ mod uv;
 mod version;
 
 pub(crate) use python::Python;
+pub(crate) use python::python_exec;
 pub(crate) use uv::Uv;
 pub(crate) use version::PythonRequest;
